feat: Criado UsuarioBuilder classe e ajustado implementações.
