+++ src/main/java/br/com/scripta_api/usuario_service/application/gateways/CustomUsuarioDetails.java
@@ -8,7 +8,9 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.List;
 
-public class CustomUsuarioDetails extends Usuario implements UserDetails {
+public class CustomUsuarioDetails
+        extends Usuario
+        implements UserDetails {
     private Usuario usuario;
 
     public CustomUsuarioDetails(Usuario usuario) {
