+++ src/main/java/br/com/scripta_api/usuario_service/repository/mapper/UsuarioMapper.java
@@ -1,6 +1,7 @@
 package br.com.scripta_api.usuario_service.repository.mapper;
 
 import br.com.scripta_api.usuario_service.application.domain.Usuario;
+import br.com.scripta_api.usuario_service.application.domain.UsuarioBuilder;
 import br.com.scripta_api.usuario_service.infra.data.UsuarioEntity;
 import org.springframework.stereotype.Component;
 
@@ -9,7 +10,7 @@ public class UsuarioMapper {
 
     public Usuario toDomain(UsuarioEntity entity) {
         if (entity == null) return null;
-        return new Usuario.builder()
+        return UsuarioBuilder.builder()
                 .id(entity.getId())
                 .nome(entity.getNome())
                 .matricula(entity.getMatricula())
